refactor: use the middleware for resetting password
